refactor: Move logic to improve user prompt and script to YamlWorkflowBuilder. Move logic to execute script to YamlExecutor from OpenRouterService. Rename OpenRouterService to OpenRouterClient. Update error catch in server endpoints and client commands
